src/requirements.txt
localstack
boto3
boto3-type-annotations
pytest
pyyaml
python-telegram

